requirements.txt
streamlit==1.41.1
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
rapidfuzz==3.11.0
praw==7.8.1
feedparser==6.0.11
pypdf==5.1.0
pymupdf==1.25.1
langchain==0.3.14
langchain-core==0.3.30
langchain-community==0.3.14
langchain-groq==0.2.3
langchain-chroma==0.2.0
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5
chromadb==0.5.23
pydantic==2.10.4
typing_extensions==4.12.2
duckduckgo-search==7.2.1